admin/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
sqlmodel>=0.0.14
redis>=4.0.0
passlib[bcrypt]>=1.7.0
loguru>=0.7.0
pymysql>=1.0.0
httpx>=0.25.0
